Implement 5 audit-derived P2 tasks: TASK-24H-1230/1231/1233/1234/1236
TASK-24H-1230: document wsprintfA's 1024-byte-write-regardless-of-buffer-
size hazard on its declaration and in docs/supported-apis.md. Docs only.

TASK-24H-1231: cast to int64_t before multiplying in CreateBitmap's pitch
computation, closing a signed-int-overflow UB path for implausibly large
widths (matches the pixel-buffer size computation's existing pattern).

TASK-24H-1233: MixerThread now captures MM_MCINOTIFY's target/id while
holding GetMidiState().mtx and posts only after releasing it, closing a
widened-critical-section/latent-lock-order issue. Added a lock-order
comment on MidiState::mtx documenting the rule for future changes.

TASK-24H-1234: removed the dead, unsynchronized g_activeTimerIds set
(written by timeSetEvent/timeKillEvent, confirmed via exhaustive grep to
never be read anywhere).

TASK-24H-1236: PeekMessageA's pendingTimers vector is now thread_local
and cleared (not freshly constructed) each call, matching StretchBlt's
existing reusable-vector pattern -- avoids a heap allocation on the
common (timer-elapsed) path after the first call.

+++ src/wingdi_bitmap.cpp
@@ -119,7 +119,12 @@ HBITMAP WINAPI CreateBitmap(int nWidth, int nHeight, UINT nPlanes, UINT nBitCoun
     bitmap->width        = nWidth;
     bitmap->height       = nHeight;
     bitmap->bitsPerPixel = 32; // store as RGBA32 internally
-    bitmap->pitch        = nWidth * 4;
+    // TASK-24H-1231: cast before multiplying, matching the pixel-buffer size
+    // computation two lines below -- a plain `nWidth * 4` int*int
+    // multiplication is undefined behavior (signed overflow) once nWidth
+    // exceeds roughly 536,870,911. Not reachable by either game's real,
+    // small, fixed bitmap dimensions; purely defensive.
+    bitmap->pitch        = static_cast<int>(static_cast<int64_t>(nWidth) * 4);
     bitmap->pixels.resize(static_cast<size_t>(nWidth) * static_cast<size_t>(nHeight) * 4u, 0);
     AdjustDiagLiveBytes(g_diagCompatBitmapPixelCapacityBytes,
                         g_diagCompatBitmapPixelCapacityHighWaterBytes,
